+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 - Support password recovery using SMTP
 - Support Dask from Jupyter notebook
 
+### Bug fixes
+- Use specific chart versions and image tags for MLFlow, Ambassador, and PostgresQL
+
 ## 0.1.0
 
 ### Features